fix(rotating_basis): F32 lab-spin path + extend regression test
`split_step_lab!` (Phase III reference path) calls `apply_ddi_step!`
directly without the Û_B wrap that `split_step_rotating!` uses. Same
Float64 dt lock as the rotating-frame DDI; needed the same boundary
conversion. Added a test_rotating_basis_f32 assertion to pin it.

Tests: 15/15 (was 14, +1 for `split_step_lab!` runnability check).

+++ src/rotating_basis_gpe.jl
@@ -553,10 +553,12 @@ function split_step_lab!(
     apply_spatial_diagonal_step!(ws, half; imaginary_time)
     apply_lab_spin_step!(ws, half, t_mid; imaginary_time)
 
-    # DDI directly in lab basis (no Û_B wrap)
+    # DDI directly in lab basis (no Û_B wrap). apply_ddi_step! locks dt
+    # to Float64 (legacy spinor solver path); convert at the boundary so
+    # F32 workspaces interop. Per-voxel array work stays at workspace T.
     if abs(ws.ddi_params.C_dd) > 1e-30
         apply_ddi_step!(ws.psi_tilde, ws.spin_matrices, ws.ddi_params,
-            ws.ddi_bufs, dt, N; imaginary_time)
+            ws.ddi_bufs, Float64(dt), N; imaginary_time)
     end
 
     apply_lab_spin_step!(ws, half, t_mid; imaginary_time)
